refactor(ui): share session MCP store actions
